requirements.txt
aioredis==1.2.0
alipay-sdk-python==3.2.53
aliyun-python-sdk-core-v3==2.11.0
aliyun-python-sdk-dysmsapi==1.0.0
amqp==2.3.2
anyjson==0.3.3
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.11.2
Automat==0.7.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
channels==2.1.6
channels-redis==2.3.2
chardet==3.0.4
constantly==15.1.0
construct==2.5.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
daphne==2.2.3
diff-match-patch==20181111
Django==2.0.7
django-celery==3.2.2
django-celery-results==1.0.4
django-ckeditor==5.6.1
django-cors-headers==2.3.0
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-filter==1.1.0
django-filters==0.2.1
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.1.0
django-js-asset==1.1.0
django-redis==4.9.0
django-rest-swagger==2.2.0
django-reversion==3.0.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
drf-haystack==1.8.3
drf-writable-nested==0.5.0
elasticsearch==2.4.1
et-xmlfile==1.0.1
fdfs-client-py==1.2.6
future==0.15.2
haystack==0.42
hiredis==0.3.0
httplib2==0.9.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itsdangerous==0.24
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
kombu==4.2.2
lxml==4.2.5
Markdown==2.6.11
MarkupSafe==1.0
msgpack==0.5.6
mutagen==1.41.0
numpy==1.15.1
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.10
pandas==0.23.4
pefile==2018.8.8
Pillow==5.2.0
pyasn1==0.4.4
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.7.2
PyHamcrest==1.9.0
PyJWT==1.6.4
PyMySQL==0.9.1
python-dateutil==2.7.3
python-ptrace==0.9.3
pytz==2018.5
PyYAML==3.13
redis==2.10.6
requests==2.19.1
rsa==4.0
simplejson==3.16.0
six==1.12.0
tablib==0.12.1
Twisted==18.9.0
txaio==18.8.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
xlrd==1.1.0
XlsxWriter==1.1.2
xlwt==1.3.0
zope.interface==4.6.0
